backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
transformers
torch
Pillow
google-generativeai
chromadb
PyPDF2
pydantic
